Refactored module calls. Moved helpers functions to a seperate module.
