+++ textworld/envs/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT license.
 
-import textworld.envs.wrappers
 from textworld.envs.zmachine.frotz import FrotzEnvironment
 from textworld.envs.glulx.git_glulx_ml import GitGlulxMLEnvironment as GlulxEnvironment
