+++ textworld/render/__init__.py
@@ -3,4 +3,4 @@
 
 
 from textworld.render.render import load_state, load_state_from_game_state, visualize
-import textworld.render.serve
+from textworld.render.serve import get_html_template
